#refactor #fix - fixing merge conflicts - changed id type to int (from long) - connected the account page to dashboard Current functionality: - log in - sign up - view acc. page - click home icon to open dashboard - view dashboard - create new project

+++ portfolio/src/main/java/nz/ac/canterbury/seng302/portfolio/model/ProjectRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ProjectRepository extends CrudRepository<Project, Long> {
+public interface ProjectRepository extends CrudRepository<Project, Integer> {
 
     public List<Project> findByProjectName(String projectName);
